Fixed warnings (and bugs found out through compile warning): 1. Changed load function name to loadExpr in flush policy and filter (due to overloaded virtual hidden warning) 2. Using elog time instead of chrono in timed flush policy (performance) 3. Fixed log record member initialization order 4. Fixed signed/unsigned mismatch in log buffer 5. Fixed member initialization order in buffered file target 6. Removed some unused local variables 7. Commented out unused static functions (may be used in the future). 8. Added missing default case for some switch-case statements. 9. Fixed bug in loading log source affinity configuration (was using wrong suffix length of log level instead of using log affinity suffix length).

+++ src/elog/src/elog_buffer.cpp
@@ -63,11 +63,14 @@ bool ELogBuffer::appendV(const char* fmt, va_list args) {
     va_copy(argsCopy, args);
     uint32_t sizeLeft = size() - m_offset;
     int res = vsnprintf(getRef() + m_offset, sizeLeft, fmt, args);
-    // return value does not include the terminating null, and number of copied characters,
+
+    // NOTE: return value does not include the terminating null, and number of copied characters,
     // including the terminating null, will not exceed size, so if res==size it means size - 1
     // characters were copied and one more terminating null, meaning one character was lost.
     // if res > size if definitely means buffer was too small, and res shows the required size
-    if (res < sizeLeft) {
+
+    // NOTE: cast to int is safe (since size is limited to ELOG_MAX_BUFFER_SIZE = 16KB)
+    if (res < (int)sizeLeft) {
         va_end(argsCopy);
         m_offset += res;
         return true;
@@ -81,7 +84,8 @@ bool ELogBuffer::appendV(const char* fmt, va_list args) {
     sizeLeft = size() - m_offset;
     res = vsnprintf(getRef() + m_offset, sizeLeft, fmt, argsCopy);
     va_end(argsCopy);
-    if (res >= sizeLeft) {
+    // NOTE: cast to int is safe (since size is limited to ELOG_MAX_BUFFER_SIZE = 16KB)
+    if (res >= (int)sizeLeft) {
         ELOG_REPORT_ERROR("Failed to format string second time");
         return false;
     }
